Handle unexpanded and expended fusion node differently.
Treat the unexpanded fusion as a normal instruction node. Only connect to the root node when the fusion is expanded. Also not connect the expanded fusion node to its operands because they will directly connect to the parameter nodes. This will make the edges less messy and easier to understand.

PiperOrigin-RevId: 782096912