Fix logic error in `MaybeAddNestedRegion`
In old logic, if an operation has `regions`, then we will add the nested region inputs inside this op. However, when we are actually processing the nested regions, we categorize them based on different dialects. If the op is not found, we won't add the regions for this op. This creates a problem.

In new logic, we use StatusOr<bool> instead of Status. bool is used to identify whether the regions are processed yet. If all dialect specific processing don't apply to it, we fallback to generic processing.

PiperOrigin-RevId: 786019818